Fix for issue #32, va_list parameters
Rather than not generating an attach_function call for any function
with va_list parameters, we're going to convert them to a :pointer
type.  This is to support that one random case where someone knows
how to correctly format a va_list in memory, and really wants to
call one of these functions.

Closes #32